requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==2.2.8
django-bootstrap4==2.2.0
django-braces==1.13.0
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-embed-video==1.1.2
django-js-asset==1.2.2
django-memcache-status==1.3
djangorestframework==3.8.2
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
python-memcached==1.59
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
